requirements.txt
google-cloud-bigquery
pandas
streamlit
plotly
db-dtypes
